pyproject.toml
[project]
name = "geopol-forecast-council"
version = "0.1.0"
description = "Lean geopolitical forecasting council with RSS+Sonar SITREP and parameterised horizons"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "click>=8.1",
    "python-dotenv>=1.0",
    "feedparser>=6.0",
    "tavily-python>=0.5",
]

[project.scripts]
forecast = "forecast.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["forecast*"]
